+++ server/middlewares/authorize.js
@@ -9,17 +9,17 @@ function authorize(allowedRoles, idRequiredRoles=[]) {
     ////////////////////////////////////////////////////////////////////////////////
     return (req, res, next) => {
         // Check if req.body.user exists
-        if (!req.body.user) {
+        if (!req.user) {
             return unauthorizedRequestResponse(res);
         }
 
         // Check if the user's role is in the allowedRoles array
-        if (!allowedRoles.includes(req.body.user[USERS_ROLE])) {
+        if (!allowedRoles.includes(req.user[USERS_ROLE])) {
             return forbiddenRequestResponse(res);
         }
 
         // Check if the user's role is in the idRequiredRoles array
-        if (idRequiredRoles.includes(req.body.user[USERS_ROLE]) && req.body.user[USERS_USER_ID] !== req.params.id) {
+        if (idRequiredRoles.includes(req.user[USERS_ROLE]) && req.user[USERS_USER_ID] !== req.params.id) {
             return forbiddenRequestResponse(res);
         }